requirements.txt
requests
pandas
numpy
tabulate
openpyxl
pytz
python-dotenv